pyproject.toml
[tool.poetry]
name = "ml_recruiters"
version = "0.1.0"
description = "Every person needs help from other professionals be it electricians, plumbers, mechanics or any other. This project is a quick and easy way for people(recruiters) to connect with skilled people who think are a good fit and hire them"
authors = ["Strategy-Tap PLC <StrategyTap@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.3"
argon2-cffi = "^20.1"
confusable_homoglyphs = "^3.2"
dj-database-url = "^0.5"
django = "^3.0"
djangorestframework = "^3.11"
django-axes = "^5.3"
django-cors-headers = "^3.2"
django-filter = "^2.2"
django-import-export = "^2.1"
django-storages = {extras = ["dropbox"], version = "^1.9"}
drf-yasg = "^1.17"
gunicorn = "^20.0"
Pillow = "^7.1"
psycopg2-binary = "^2.8"
python-decouple = "^3.3"
pwned-passwords-django = "^1.4"
whitenoise = {extras = ["brotli"], version = "^5.0"}
loguru = "^0.4.1"
drf-typed-views = "^0.2.1"
pydantic = "^1.5"
dj-rest-auth = {extras = ["with_social"], version = "^1.0"}
djangorestframework-simplejwt = "^4.4.0"
django-simple-history = "^2.10.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6"
black = "^19.10b0"
coverage = "^5.0"
django-coverage-plugin = "^1.8"
django-debug-toolbar = "^2.2"
factory-boy = "^2.12"
hypothesis = {extras = ["django"], version = "^5.10"}
isort = "^4.3"
mypy = "^0.770"
mkdocs = "^1.1"
mkdocs-material = "^5"
pre-commit = "^2.3"
pytest = "^5.4"
pytest-cov = "^2.8"
pytest-django = "^3.9"
pytest-sugar = "^0.9"
radon = "^4.1"
safety = "^1.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
